requirements.txt
PyQt6>=6.0
python-dotenv>=1.0.0
pymodbus>=2.5.3